Pacient Tests

+++ src/modules/patients/useCases/deletePatient/DeletePatientUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 
 import { IPatientsRepository } from "@modules/patients/repositories/IPatientsRepository";
+import { AppError } from "@shared/errors/AppError";
 
 @injectable()
 class DeletePatientUseCase {
@@ -10,6 +11,11 @@ class DeletePatientUseCase {
   ) {}
 
   public async execute(id: string): Promise<void> {
+    const patient = await this.patientsRepository.findById(id);
+
+    if (!patient) {
+      throw new AppError("Patient doesn't exists", 404);
+    }
     await this.patientsRepository.delete(id);
   }
 }
